Expose context to middleware
A recent upstream change bumped Work to go-redis/v8, which uses a
context everywhere. Despite this, the context wasn't available to the
middleware, which meant we couldn't pass execution-specific
information to the actual handler.

This commit effectively moves everything from HandlerFunc to
ContextHandlerFunc, passing the context to the middleware and
underlying operations.